# lib/api-zod/src/generated/types/debugSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Dev Debugger API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DebugSessionSeverity } from "./debugSessionSeverity";

export interface DebugSession {
  id: string;
  rootCause: string;
  fix: string;
  explanation: string;
  correctedCode: string;
  proTip: string;
  language: string;
  severity: DebugSessionSeverity;
  /** @nullable */
  tokensUsed?: number | null;
  cached: boolean;
  /** @nullable */
  durationMs?: number | null;
  createdAt: string;
  /** @nullable */
  rawCode?: string | null;
  /** @nullable */
  rawError?: string | null;
  mode?: string;
  /** @nullable */
  creditsRemaining?: number | null;
}