# lib/api-zod/src/generated/types/languageStat.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Dev Debugger API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LanguageStat {
  language: string;
  count: number;
}