# lib/api-zod/src/generated/types/publicUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Smart Dev Debugger API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PublicUserAuthMethod } from "./publicUserAuthMethod";

export interface PublicUser {
  id: string;
  fullName: string;
  email: string;
  /** @nullable */
  avatarUrl: string | null;
  credits: number;
  authMethod: PublicUserAuthMethod;
  createdAt: string;
}